Code refactor

+++ vsix_downloader/downloader.py
@@ -1,5 +1,4 @@
 import asyncio
-from pathlib import Path
 from typing import Collection
 
 import aiofile
@@ -7,7 +6,7 @@ import httpx
 from tqdm.asyncio import tqdm
 
 from .api import VSCodeExtensionAPI
-from .data import VSCodeExtension, VSCodeExtensionVersion
+from .data import VSCodeExtension, VSCodeExtensionVersion, DownloadOptions, VersionFilterOptions
 from .utils import download_file, get_latest_extension_version
 
 
@@ -16,42 +15,48 @@ async def _run_download_task(
         ext_name: str,
         extension: VSCodeExtension,
         version: VSCodeExtensionVersion,
-        target_dir: Path,
-        temp_dir: Path,
+        download_options: DownloadOptions
 ) -> None:
-    extension_dir = target_dir.joinpath(ext_name)
+    if download_options.flatten_dir:
+        extension_dir = download_options.target_dir
+    else:
+        extension_dir = download_options.target_dir.joinpath(ext_name)
     extension_dir.mkdir(parents=True, exist_ok=True)
-    platform_tag = version.target_platform if version.target_platform else "all"
-    extension_meta_name = f"{ext_name}_{version.version}_{platform_tag}.json"
-    async with aiofile.async_open(extension_dir.joinpath(extension_meta_name), "w", encoding="utf-8") as f:
-        download_meta = VSCodeExtension(
-            extension_id=extension.extension_id,
-            extension_name=extension.extension_name,
-            display_name=extension.display_name,
-            versions=(version,)
+
+    if version.target_platform:
+        download_name = f"{ext_name}-{version.version}@{version.target_platform}"
+    else:
+        download_name = f"{ext_name}-{version.version}"
+
+    meta_data_path = extension_dir.joinpath(f"{download_name}.json")
+    if not download_options.no_metadata and (not download_options.skip_if_exists or not meta_data_path.exists()):
+        async with aiofile.async_open(meta_data_path, "w", encoding="utf-8") as f:
+            download_meta = VSCodeExtension(
+                extension_id=extension.extension_id,
+                extension_name=extension.extension_name,
+                display_name=extension.display_name,
+                versions=(version,)
+            )
+            await f.write(download_meta.to_json(indent=2, ensure_ascii=True))
+    if not download_options.skip_if_exists or not extension_dir.joinpath(f"{download_name}.vsix").exists():
+        await download_file(
+            client,
+            version.package_url,
+            extension_dir,
+            f"{download_name}.vsix",
+            download_options.temp_dir
         )
-        await f.write(download_meta.to_json(indent=2, ensure_ascii=True))
-    await download_file(
-        client,
-        version.package_url,
-        extension_dir,
-        f"{ext_name}_{version.version}_{platform_tag}.vsix",
-        temp_dir
-    )
 
 
 async def download_extensions(
-        target_dir: Path,
-        temp_dir: Path,
         ext_names: Collection[str],
-        target_platform: str | None,
-        vscode_version: str | None,
-        include_prerelease: bool = False,
+        download_options: DownloadOptions,
+        version_filter_options: VersionFilterOptions
 ):
     if len(ext_names) == 0:
         return
 
-    target_dir.mkdir(parents=True, exist_ok=True)
+    download_options.target_dir.mkdir(parents=True, exist_ok=True)
     async with httpx.AsyncClient(timeout=httpx.Timeout(30.0)) as client:
         api = VSCodeExtensionAPI(client)
         extensions: dict[str, VSCodeExtension] = await api.get_extensions(ext_names)
@@ -61,7 +66,7 @@ async def download_extensions(
 
         download_tasks = []
         for ext_name, extension in extensions.items():
-            version = get_latest_extension_version(extension, target_platform, vscode_version, include_prerelease)
+            version = get_latest_extension_version(extension, version_filter_options)
             if version is not None:
                 download_tasks.append(
                     asyncio.create_task(
@@ -70,8 +75,7 @@ async def download_extensions(
                             ext_name=ext_name,
                             extension=extension,
                             version=version,
-                            target_dir=target_dir,
-                            temp_dir=temp_dir
+                            download_options=download_options
                         )
                     )
                 )
